# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ContaraNAS API",
    "description": "Backend API for ContaraNAS",
    "version": "0.1.0"
  },
  "paths": {
    "/api/modules": {
      "get": {
        "tags": [
          "commands"
        ],
        "summary": "List Modules",
        "description": "List all registered modules with their current state",
        "operationId": "list_modules_api_modules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/marketplace/modules": {
      "get": {
        "tags": [
          "marketplace"
        ],
        "summary": "List Modules",
        "description": "List all available modules from marketplace",
        "operationId": "list_modules_api_marketplace_modules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/marketplace/modules/{module_id}": {
      "get": {
        "tags": [
          "marketplace"
        ],
        "summary": "Get Module",
        "description": "Get detailed information about a specific module",
        "operationId": "get_module_api_marketplace_modules__module_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "module_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Module Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/marketplace/refresh": {
      "post": {
        "tags": [
          "marketplace"
        ],
        "summary": "Refresh Registry",
        "description": "Force refresh the marketplace registry cache",
        "operationId": "refresh_registry_api_marketplace_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModuleInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "initialized": {
            "type": "boolean",
            "title": "Initialized"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "removable": {
            "type": "boolean",
            "title": "Removable"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "system_deps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "System Deps"
          }
        },
        "type": "object",
        "required": [
          "name",
          "display_name",
          "enabled",
          "initialized",
          "source",
          "removable",
          "version",
          "system_deps"
        ],
        "title": "ModuleInfo",
        "description": "Information about a module"
      },
      "ModuleListResponse": {
        "properties": {
          "modules": {
            "items": {
              "$ref": "#/components/schemas/ModuleInfo"
            },
            "type": "array",
            "title": "Modules"
          }
        },
        "type": "object",
        "required": [
          "modules"
        ],
        "title": "ModuleListResponse",
        "description": "Response listing all modules"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}